Make ngrok download modular (#160)
* Make ngrok download modular

* Add new file to package.json + README update
