+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "ngrok",
-  "version": "3.2.5",
+  "version": "3.2.6",
   "description": "node wrapper for ngrok",
   "main": "index.js",
   "types": "ngrok.d.ts",
@@ -11,6 +11,7 @@
   },
   "files": [
     "index.js",
+    "download.js",
     "postinstall.js",
     "process.js",
     "bin/ngrok",
